+++ infobeans-home/api/routes/routes.js
@@ -1,46 +1,8 @@
 const express = require("express");
 const jwt = require("jsonwebtoken");
-const multer = require("multer");
 
-const transporter = require("../config/mail");
 const mailer = require("../controller/mailControl");
-
-/* How we want to store our multer file */
-
-// binary data
-// const upload = multer({ dest: "./uploads/" });
-
-// disk storage
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, "uploads");
-  },
-  filename: function (req, file, cb) {
-    cb(null, `${Date.now()}-${file.originalname}`);
-  },
-});
-
-// only store certain file types
-const fileFilter = (req, file, cb) => {
-  if (
-    file.mimetype === "image/jpeg" ||
-    file.mimetype === "image/jpg" ||
-    file.mimetype === "image/png" ||
-    file.mimetype === "application/pdf"
-  ) {
-    cb(null, true);
-  } else {
-    cb(null, false);
-  }
-};
-
-const upload = multer({
-  storage: storage,
-  limits: {
-    fileSize: 1024 * 1024 * 5, // max 5 MB
-  },
-  fileFilter,
-});
+const upload = require("../config/multer");
 
 const Users = require("../model/userSchema");
 const Jobs = require("../model/jobSchema");
